+++ src/components/User/Assets/Issuance/IssuanceDetails.vue
@@ -129,7 +129,7 @@ export default {
       try {
         this.issuance = await api.requests.get(id)
       } catch (error) {
-        this.$store.dispatch('SET_ERROR', 'Cannot load issuance request list.')
+        ErrorHandler.processWithoutFeedback(error)
       }
     },
     async fulfill (issuance) {